[FEATURE] 네이버 로그아웃/탈퇴 구현
- 마스터 계정 예외처리

+++ app/src/main/java/com/gritbus/hipchon/ui/my/viewmodel/MySettingViewModel.kt
@@ -31,6 +31,14 @@ class MySettingViewModel @Inject constructor(
     }
 
     fun leaveUser() {
+        if (UserData.userLoginId == "masterId") {
+            userRepository.setAutoLoginId(null)
+            userRepository.setAutoLoginPlatform(null)
+
+            _isLeaveSuccess.value =
+                userRepository.getAutoLoginId() == null && userRepository.getAutoLoginPlatform() == null
+            return
+        }
         viewModelScope.launch {
             userRepository.deleteUserData(UserData.platform, UserData.userLoginId)
                 .onSuccess {
